fix: share database transfer api errors
